services/python-worker-v2/requirements.txt
grpcio
grpcio-tools
protobuf
vietocr
torch
torchvision
